small cleanup and added validations
