+++ modules/channel-web/src/views/lite/main.jsx
@@ -5,8 +5,10 @@ import addMilliseconds from 'date-fns/add_milliseconds'
 import isBefore from 'date-fns/is_before'
 import queryString from 'query-string'
 import ms from 'ms'
-
+import ChatIcon from './icons/Chat'
+import CloseIcon from './icons/CloseChat'
 import Container from './components/Container'
+import { downloadFile } from './utils'
 
 const _values = obj => Object.keys(obj).map(x => obj[x])
 
@@ -270,19 +272,23 @@ export default class Web extends React.Component {
     }
   }
 
-  fetchData = () => {
-    return this.fetchBotInfo()
-      .then(this.fetchConversations)
-      .then(this.fetchCurrentConversation)
-      .then(() => {
-        const locale = navigator.language || navigator.userLanguage
-        this.handleSendData({
-          type: 'visit',
-          text: 'User visit',
-          timezone: new Date().getTimezoneOffset() / 60,
-          language: locale && locale.substring(0, locale.indexOf('-'))
-        }).catch(this.checkForExpiredExternalToken)
-      })
+  fetchData = async () => {
+    try {
+      await this.fetchBotInfo()
+        .then(this.fetchConversations)
+        .then(this.fetchCurrentConversation)
+    } catch (err) {
+      console.log('Error while fetching data, creating new convo...', err)
+      await this.createConversation()
+    }
+
+    const locale = navigator.language || navigator.userLanguage
+    this.handleSendData({
+      type: 'visit',
+      text: 'User visit',
+      timezone: new Date().getTimezoneOffset() / 60,
+      language: locale && locale.substring(0, locale.indexOf('-'))
+    }).catch(this.checkForExpiredExternalToken)
   }
 
   fetchConversations = () => {
@@ -490,34 +496,6 @@ export default class Web extends React.Component {
     return this.props.bp.axios.post(url, {}, this.axiosConfig).then()
   }
 
-  renderOpenIcon() {
-    return (
-      <svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M4.583 14.894l-3.256 3.78c-.7.813-1.26.598-1.25-.46a10689.413 10689.413 0 0 1 .035-4.775V4.816a3.89 3.89 0 0 1 3.88-3.89h12.064a3.885 3.885 0 0 1 3.882 3.89v6.185a3.89 3.89 0 0 1-3.882 3.89H4.583z"
-          fill="#FFF"
-          fillRule="evenodd"
-        />
-      </svg>
-    )
-  }
-
-  renderCloseIcon() {
-    return (
-      <svg width="17" height="17" viewBox="0 0 17 17" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M16.726 15.402c.365.366.365.96 0 1.324-.178.178-.416.274-.663.274-.246 0-.484-.096-.663-.274L8.323 9.648h.353L1.6 16.726c-.177.178-.416.274-.663.274-.246 0-.484-.096-.663-.274-.365-.365-.365-.958 0-1.324L7.35 8.324v.35L.275 1.6C-.09 1.233-.09.64.274.274c.367-.365.96-.365 1.326 0l7.076 7.078h-.353L15.4.274c.366-.365.96-.365 1.326 0 .365.366.365.958 0 1.324L9.65 8.675v-.35l7.076 7.077z"
-          fill="#FFF"
-          fillRule="evenodd"
-        />
-      </svg>
-    )
-  }
-
-  renderUncountMessages() {
-    return <span className={'bpw-floating-button-unread'}>{this.state.unreadCount}</span>
-  }
-
   renderWidget() {
     if (this.state.isButtonHidden) {
       return null
@@ -529,32 +507,20 @@ export default class Web extends React.Component {
         })}
         onClick={this.handleButtonClicked}
       >
-        <i>{this.state.view === 'convo' ? this.renderCloseIcon() : this.renderOpenIcon()}</i>
-        {this.state.unreadCount > 0 ? this.renderUncountMessages() : null}
+        {this.state.view === 'convo' ? <CloseIcon /> : <ChatIcon />}
+        {this.state.unreadCount > 0 && <span className={'bpw-floating-button-unread'}>{this.state.unreadCount}</span>}
       </button>
     )
   }
 
-  createConversation = () => {
+  createConversation = async () => {
     const userId = window.__BP_VISITOR_ID
     const url = `/mod/channel-web/conversations/${userId}/new`
+    const { data } = await this.props.bp.axios.post(url, {}, this.axiosConfig)
 
-    // TODO here we might we might want switch convo with the newly created conversation (need to return the convo ID in BE)
-    return this.props.bp.axios.post(url, {}, this.axiosConfig).then(this.fetchConversations)
-  }
-
-  downloadFile(name, blob) {
-    const url = window.URL.createObjectURL(blob)
-    const link = document.createElement('a')
-
-    link.href = url
-    link.setAttribute('download', name)
-
-    document.body.appendChild(link)
-    link.click()
-
-    document.body.removeChild(link)
-    window.URL.revokeObjectURL(url)
+    this.fetchConversations().then(() => {
+      this.handleSwitchConvo(data.convoId)
+    })
   }
 
   downloadConversation = async () => {
@@ -563,7 +529,7 @@ export default class Web extends React.Component {
     const file = (await this.props.bp.axios.get(url, this.axiosConfig)).data
     const blobFile = new Blob([file.txt])
 
-    this.downloadFile(file.name, blobFile)
+    downloadFile(file.name, blobFile)
   }
 
   renderSide() {
