+++ modules/channel-web/src/views/lite/utils.js
@@ -22,3 +22,17 @@ export const asyncDebounce = async timeMs => {
     }
   }
 }
+
+export const downloadFile = (name, blob) => {
+  const url = window.URL.createObjectURL(blob)
+  const link = document.createElement('a')
+
+  link.href = url
+  link.setAttribute('download', name)
+
+  document.body.appendChild(link)
+  link.click()
+
+  document.body.removeChild(link)
+  window.URL.revokeObjectURL(url)
+}
